move monkey-patching to test dir
