+++ ml-agents/mlagents/tf_utils/tf.py
@@ -28,50 +28,3 @@ def set_warnings_enabled(is_enabled: bool) -> None:
     """
     level = tf_logging.WARN if is_enabled else tf_logging.ERROR
     tf_logging.set_verbosity(level)
-
-
-# HACK numpy here for now
-import numpy as np
-import traceback
-
-# TODO condition on env variable
-__old_np_array = np.array
-__old_np_zeros = np.zeros
-__old_np_ones = np.ones
-
-
-def _check_no_float64(arr, kwargs_dtype):
-    if arr.dtype == np.float64:
-        tb = traceback.extract_stack()
-        # tb[-1] in the stack is this function.
-        # tb[-2] is the wrapper function, e.g. np_array_no_float64
-        # we want the calling function, so use tb[-3]
-        filename = tb[-3].filename
-        # HACK - only raise if this came from mlagents code, not tensorflow
-        if (
-            "ml-agents/mlagents" in filename or "ml-agents-envs/mlagents" in filename
-        ) and "tensorflow_to_barracuda.py" not in filename:
-            raise ValueError(f"dtype={kwargs_dtype}")
-
-
-def np_array_no_float64(*args, **kwargs):
-    res = __old_np_array(*args, **kwargs)
-    _check_no_float64(res, kwargs.get("dtype"))
-    return res
-
-
-def np_zeros_no_float64(*args, **kwargs):
-    res = __old_np_zeros(*args, **kwargs)
-    _check_no_float64(res, kwargs.get("dtype"))
-    return res
-
-
-def np_ones_no_float64(*args, **kwargs):
-    res = __old_np_ones(*args, **kwargs)
-    _check_no_float64(res, kwargs.get("dtype"))
-    return res
-
-
-np.array = np_array_no_float64
-np.zeros = np_zeros_no_float64
-np.ones = np_ones_no_float64
